backend/ciudadanos/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pydantic
python-dotenv
python-jose
pydantic[email]
aio_pika